1.订单拣货类型判断修改 2.cne创建包裹失败后保存失败原因

+++ oms/src/main/java/com/upedge/oms/modules/order/dao/OrderDao.java
@@ -394,6 +394,6 @@ public interface OrderDao {
 
     List<OrderItemQuantityVo> selectOrderItemQuantities(OrderItemQuantityDto orderItemQuantityDto);
 
-    int updateOrderPickState(@Param("orderIds") List<Long> orderIds, @Param("state") Integer state, @Param("waveNo") Long waveNo);
+    int updateOrderPickState(@Param("orderIds") List<Long> orderIds, @Param("state") Integer state, @Param("waveNo") Integer waveNo);
 
 }
