+++ oms/src/main/java/com/upedge/oms/modules/pack/entity/OrderPackage.java
@@ -37,7 +37,7 @@ public class OrderPackage{
 	/**
 	 * 
 	 */
-    private Long waveNo;
+    private Integer waveNo;
 	/**
 	 * 真实追踪号
 	 */
